+++ src/pages/Exchange/MainArea/Trade/Chart/Styles.ts
@@ -1,0 +1,3 @@
+import styled from 'styled-components';
+
+export const Container = styled.div``;
