+++ src/pages/Exchange/MainArea/Trade/OrderBook/Styles.ts
@@ -14,6 +14,7 @@ const sellMixin = css`
 export const Container = styled.div`
   border-radius: 4px;
   border: 1px solid ${colors.border};
+  margin-top: 16px;
   padding: 16px;
 `;
