+++ src/pages/Exchange/MainArea/Trade/OrderTools/Buy/Styles.ts
