+++ src/pages/Exchange/MainArea/Trade/OrderTools/Sell/Styles.ts
