+++ src/pages/Exchange/MainArea/Trade/OrderTools/Styles.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+import UTabs from 'components/Tabs';
+
+export const Container = styled.div``;
+
+export const TabContent = styled.div`
+  margin-top: 16px;
+`;
+
+export const Tabs = styled(UTabs)``;
