+++ src/pages/Exchange/MainArea/Trade/Styles.ts
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
 
-import UTabs from 'components/Tabs';
 import {breakpoints} from 'styles';
 
 export const Container = styled.div`
@@ -14,13 +13,9 @@ export const Grid = styled.div`
   grid-template-columns: min-content auto;
   margin-top: 12px;
 
-  @media (max-width: ${breakpoints.mobile}) {
+  @media (max-width: ${breakpoints.tablet}) {
     grid-template-columns: auto;
   }
 `;
 
-export const TabContent = styled.div``;
-
-export const Tabs = styled(UTabs)`
-  margin-top: 16px;
-`;
+export const OrderHistory = styled.div``;
